Created Hunter's Sniff and Animal Corpses
Sniff:
Notes: particles, sounds, variable values are not there.
temporary variables and particles are being used as they are dependant
on the functionality.

Animal Corpses:
"npc_creep_corpse"
Notes: This unit is spawned whenever a creep (animal) is killed, and is
used to store information about the killer, and also to be targeted by
Sniff. Corpses have decay timers which is currently set to 3 minutes,
can be modified in scripts/vscripts/units/creeps/corpse.lua
The corpse model is set to "bonds_of_madness.vmdl" as a temporary model.