Fixed #148
Due to the fact that spears aren't consumed instantly I had to very
cleverly add in a timer. Due to this theres a small window where someone
can drop the item with 0 charges (which it can still be used) from the
tower. I inserted a comment where we need to make it so the building
cant preform any dropping functions at that time.