Development (#374)
* Convert addon_english.txt from UTF-16 to UTF-8

* Fix misspellings, grammar, and inconsistencies

* Add error message for empty bushes

* Make smoke house eject unstackable smoked meat

The smoke house was consuming cooked meat but giving nothing in return
when smoked meat couldn't be stacked. This makes the smoke house eject
a smoked meat item on the ground.
An alternative solution would be to check for cooked meat, attempt to
give the smoke house a smoked meat, and only remove a cooked meat if
it succeeded.

* Make crafted items stack

* Attempt to fix bug where creeps become trade ships

By making the variable local, it will not accidently exist outside the
scope of the SpawnBoat() function. This may be the cause for the bug.

* Make smoke hut cure diseased meat

* Add filter for sleep buildings

Prevents sleeping from being queued when cast out of range, and only
allows sleep to be cast on self. Ideally, you'd want to move the player
slightly into range instead of erroring when they're out of range.

* Improve filter error messages

* Fix Smoke Meat tooltip